+++ cr_zkteco_biometric_integration/models/__init__.py
@@ -3,6 +3,7 @@
 
 from . import biometric_device_config
 from . import biometric_attendance_log
+from . import biometric_attendance_alert_log
 from . import hr_employee_extend
 from . import biometric_dashboard
 from . import biometric_device_command